requirements.txt
aiofiles==23.2.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
bidict==0.23.1
billiard==4.2.0
blinker==1.7.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==4.4.2
distro==1.9.0
fastapi==0.110.2
Flask==3.0.3
Flask-Testing==0.8.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-texttospeech==2.16.3
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gTTS==2.5.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.7
MarkupSafe==2.1.5
moviepy==1.0.3
numpy==1.26.4
openai==1.19.0
packaging==24.0
pillow==10.3.0
pluggy==1.4.0
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-engineio==4.9.0
python-socketio==5.11.2
pyttsx3==2.90
redis==5.0.3
requests==2.31.0
rsa==4.9
setuptools==69.5.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
wsproto==1.2.0
